--- rust_cpsat/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "rust_cpsat"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-cp_sat = { git = "https://github.com/bc-ross/cp_sat.git", rev = "2ba15818232d73fb5169fc806f57b7c96aed8bcb" }
-serde = { version = "^1.0.219", features = ["derive"] }
-anyhow = "^1.0.98"
-savefile = { version = "0.18.7", features = ["derive", "serde_derive"] }
-savefile-derive = "0.18.7"
-lazy_static = "1.5.0"
