--- rust_cpsat/src/model.rs
@@ -1,7 +1,0 @@
-//! Model building and constraint logic for the course scheduling solver.
-
-pub use crate::model_context::{Course, ModelBuilderContext, build_model_pipeline};
-pub use crate::model_courses::*;
-pub use crate::model_geneds::*;
-pub use crate::model_prereqs::*;
-pub use crate::model_semester::*;
