+++ src/model/mod.rs
@@ -1,0 +1,16 @@
+//! Model building and constraint logic for the course scheduling solver.
+
+mod context;
+mod courses;
+mod geneds;
+mod prereqs;
+mod semester;
+mod two_stage_schedule;
+
+use context::{Course, ModelBuilderContext, build_model_pipeline};
+use courses::*;
+use geneds::*;
+use prereqs::*;
+use semester::*;
+
+pub use two_stage_schedule::two_stage_lex_schedule;
